CLI#PLAY changes + supporting edits
## Discussion of methods in #play

### load_yaml
Let’s us manage code and copy separately.

### set_players_and_days
Already feels like a temporary stop in order to thin out #play.

### @gamemaster.start_game
I like the idea of letting the playing be directed by the gamemaster

### end_game
A way for us to reconcile and determine the winner.

## play_ls.rb
Not sure the argv is relevant here.

## game.rb
Moved some of the content guts out of #play